made Emitter an interface to allow custom emitters
